requirements.txt
# Core Dependencies
httpx>=0.27.0
websockets>=12.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database Drivers
aiosqlite>=0.19.0
psycopg2-binary>=2.9.9

# Security & Encryption
cryptography>=41.0.0

# Logging & Monitoring
loguru>=0.7.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Optional: Web API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Optional: CLI
click>=8.1.0
rich>=13.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
